Move account implementations and auths from rpc server to js
