+++ yarn-project/aztec-rpc/src/index.ts
@@ -4,8 +4,6 @@ export * from './aztec_rpc_server/index.js';
 export * from './aztec_rpc_http/index.js';
 export * from './tx/index.js';
 export * from './crypto/types.js';
-export * from './auth/index.js';
-export * from './account_impl/index.js';
 
 export { Tx, TxHash, PartialContractAddress } from '@aztec/types';
