+++ yarn-project/aztec.js/src/account_impl/account_contract.ts
@@ -6,9 +6,10 @@ import { PublicKey } from '@aztec/key-store';
 import { ExecutionRequest, PartialContractAddress, TxExecutionRequest } from '@aztec/types';
 import partition from 'lodash.partition';
 import times from 'lodash.times';
-import { TxAuthProvider, generateFunctionSelector } from '../index.js';
+import { generateFunctionSelector } from '../index.js';
 import { AccountImplementation } from './index.js';
 import { ContractAbi } from '@aztec/foundation/abi';
+import { TxAuthProvider } from '../auth/index.js';
 
 /**
  * Account backed by an account contract
