+++ yarn-project/aztec.js/src/account_impl/index.ts
