+++ yarn-project/aztec.js/src/auth/index.ts
